perf(web): reduce initial bundle payload
