+++ apps/web/package.json
@@ -20,16 +20,11 @@
     "@hookform/resolvers": "^5.4.0",
     "@hugeicons/core-free-icons": "^4.2.2",
     "@hugeicons/react": "^1.1.9",
-    "@tailwindcss/vite": "^4.3.2",
-    "@tanstack/react-devtools": "^0.10.8",
     "@tanstack/react-query": "^5.101.2",
     "@tanstack/react-router": "^1.170.17",
-    "@tanstack/react-router-devtools": "^1.167.0",
     "@tanstack/react-table": "^8.21.3",
     "@tanstack/react-virtual": "^3.14.6",
-    "@tanstack/router-plugin": "^1.168.19",
     "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "cnfast": "^0.0.8",
     "next-themes": "^0.4.6",
@@ -37,15 +32,16 @@
     "react-dom": "^19.2.7",
     "react-hook-form": "^7.81.0",
     "react-markdown": "^10.1.0",
-    "shadcn": "^4.13.0",
     "sonner": "^2.0.7",
-    "tailwind-merge": "^3.6.0",
-    "tailwindcss": "^4.3.2",
     "zod": "^4.4.3",
     "zustand": "^5.0.14"
   },
   "devDependencies": {
     "@tanstack/devtools-vite": "^0.8.1",
+    "@tailwindcss/vite": "^4.3.2",
+    "@tanstack/react-devtools": "^0.10.8",
+    "@tanstack/react-router-devtools": "^1.167.0",
+    "@tanstack/router-plugin": "^1.168.19",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/react": "^16.3.2",
     "@types/node": "^26.1.1",
@@ -57,6 +53,8 @@
     "oxlint": "1.73.0",
     "oxlint-tsgolint": "0.24.0",
     "prettier": "^3.9.5",
+    "shadcn": "^4.13.0",
+    "tailwindcss": "^4.3.2",
     "tw-animate-css": "^1.4.0",
     "typescript": "7.0.2",
     "vite": "^8.1.4",
