+++ apps/web/src/components/build-details/build-detail-route.tsx
@@ -1,8 +1,30 @@
+import { Suspense, lazy } from 'react'
 import { useParams, useSearch } from '@tanstack/react-router'
 
-import { ArtifactInstallPage } from './artifact-install-page'
-import { BuildDetailPage } from './build-detail-page'
+import PageLayout from '@/components/page-layout'
+import { Skeleton } from '@/components/ui/skeleton'
 import { useAuthStore } from '@/stores/auth-store'
+
+const ArtifactInstallPage = lazy(() =>
+  import('./artifact-install-page').then((module) => ({
+    default: module.ArtifactInstallPage,
+  })),
+)
+const BuildDetailPage = lazy(() =>
+  import('./build-detail-page').then((module) => ({
+    default: module.BuildDetailPage,
+  })),
+)
+
+function BuildDetailFallback() {
+  return (
+    <PageLayout width="full">
+      <Skeleton className="h-8 w-56" />
+      <Skeleton className="h-24 w-full" />
+      <Skeleton className="h-64 w-full" />
+    </PageLayout>
+  )
+}
 
 export function BuildDetailRoute() {
   const { buildId } = useParams({ from: '/builds/$buildId' })
@@ -10,12 +32,18 @@ export function BuildDetailRoute() {
   const isQaViewer = useAuthStore((state) => state.user?.role === 'qa_viewer')
   if (artifactId || isQaViewer) {
     return (
-      <ArtifactInstallPage
-        key={`${buildId}:${artifactId ?? 'auto'}`}
-        buildId={buildId}
-        artifactId={artifactId}
-      />
+      <Suspense fallback={<BuildDetailFallback />}>
+        <ArtifactInstallPage
+          key={`${buildId}:${artifactId ?? 'auto'}`}
+          buildId={buildId}
+          artifactId={artifactId}
+        />
+      </Suspense>
     )
   }
-  return <BuildDetailPage key={buildId} buildId={buildId} />
+  return (
+    <Suspense fallback={<BuildDetailFallback />}>
+      <BuildDetailPage key={buildId} buildId={buildId} />
+    </Suspense>
+  )
 }
