+++ apps/web/src/routes/__root.tsx
@@ -29,12 +29,14 @@ import { useAuthStore } from '@/stores/auth-store'
 import { useUiStore } from '@/stores/ui-store'
 import { useInstanceStore } from '@/stores/instance-store'
 import { isDemoMode } from '@/lib/demo-mode'
+import { useWindowEvent } from '@/hooks/use-window-event'
 import {
   RootErrorBoundary,
   RootNotFound,
 } from '@/components/root-route-boundaries'
 
-const CommandPalette = lazy(() => import('@/components/command-palette'))
+const loadCommandPalette = () => import('@/components/command-palette')
+const CommandPalette = lazy(loadCommandPalette)
 
 const DevTools = import.meta.env.DEV
   ? lazy(() =>
@@ -105,6 +107,8 @@ function RootLayout() {
   const authToken = useAuthStore((s) => s.token)
   const authUser = useAuthStore((s) => s.user)
   const openCommandPalette = useUiStore((s) => s.setCommandPaletteOpen)
+  const commandPaletteOpen = useUiStore((s) => s.commandPaletteOpen)
+  const toggleCommandPalette = useUiStore((s) => s.toggleCommandPalette)
   const sidebarOpen = useUiStore((s) => s.sidebarOpen)
   const setSidebarOpen = useUiStore((s) => s.setSidebarOpen)
 
@@ -118,6 +122,18 @@ function RootLayout() {
   const showQaChrome = showAppChrome && authUser.role === 'qa_viewer'
 
   useSessionMonitor()
+  useWindowEvent('keydown', (event) => {
+    if (
+      showAppChrome &&
+      !showQaChrome &&
+      (event.metaKey || event.ctrlKey) &&
+      event.key === 'k'
+    ) {
+      event.preventDefault()
+      void loadCommandPalette()
+      toggleCommandPalette()
+    }
+  })
 
   return (
     <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
@@ -146,6 +162,8 @@ function RootLayout() {
                   <button
                     type="button"
                     aria-label="Search"
+                    onMouseEnter={() => void loadCommandPalette()}
+                    onFocus={() => void loadCommandPalette()}
                     onClick={() => openCommandPalette(true)}
                     className="inline-flex h-8 w-8 items-center justify-center gap-2 rounded-sm border bg-muted/50 px-0 text-xs text-muted-foreground transition-colors hover:bg-accent hover:text-accent-foreground sm:w-48 sm:justify-between sm:px-3"
                   >
@@ -170,7 +188,7 @@ function RootLayout() {
           </div>
         )}
         <Toaster />
-        {showAppChrome && !showQaChrome ? (
+        {showAppChrome && !showQaChrome && commandPaletteOpen ? (
           <Suspense fallback={null}>
             <CommandPalette />
           </Suspense>
