+++ apps/web/src/routes/setup/index.lazy.tsx
@@ -1,0 +1,138 @@
+import { createLazyFileRoute, useNavigate } from '@tanstack/react-router'
+import { useForm } from 'react-hook-form'
+import { zodResolver } from '@hookform/resolvers/zod'
+import z from 'zod'
+import { Button } from '@/components/ui/button'
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from '@/components/ui/form'
+import { Input } from '@/components/ui/input'
+import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
+import { useSetupStatus, useVerifyBootstrapToken } from '@/hooks/use-setup'
+import { useSetupStore } from '@/stores/setup-store'
+import { getApiErrorMessage } from '@/lib/api'
+import { PageMeta } from '@/lib/seo'
+import { useBootstrapStepTransition } from '@/hooks/use-setup-route-transitions'
+
+const bootstrapTokenSchema = z.object({
+  token: z.string().min(1, 'Bootstrap token is required'),
+})
+
+type BootstrapTokenForm = z.infer<typeof bootstrapTokenSchema>
+
+export const Route = createLazyFileRoute('/setup/')({
+  component: BootstrapTokenStep,
+  errorComponent: BootstrapTokenError,
+})
+
+function BootstrapTokenError({ error }: { error: Error }) {
+  return (
+    <div className="space-y-4">
+      <Alert variant="destructive">
+        <AlertTitle>Something went wrong</AlertTitle>
+        <AlertDescription>{error.message}</AlertDescription>
+      </Alert>
+    </div>
+  )
+}
+
+function BootstrapTokenStep() {
+  const navigate = useNavigate()
+  const sessionToken = useSetupStore((s) => s.sessionToken)
+  const setSessionToken = useSetupStore((s) => s.setSessionToken)
+  const setSessionExpiresAt = useSetupStore((s) => s.setSessionExpiresAt)
+  const setCurrentStep = useSetupStore((s) => s.setCurrentStep)
+  const verifyMutation = useVerifyBootstrapToken()
+  const { data: status } = useSetupStatus()
+
+  const form = useForm<BootstrapTokenForm>({
+    resolver: zodResolver(bootstrapTokenSchema),
+    defaultValues: { token: '' },
+    mode: 'onBlur',
+  })
+
+  useBootstrapStepTransition(status, sessionToken)
+
+  const errorMessage = verifyMutation.error
+    ? getApiErrorMessage(verifyMutation.error, {
+        token_expired:
+          'Bootstrap token has expired. Generate a new one with the CLI.',
+        token_consumed:
+          'Bootstrap token has already been used. Generate a new one with the CLI.',
+        invalid_token:
+          'Invalid bootstrap token. Check the value and try again.',
+      })
+    : null
+
+  function handleFormSubmit(data: BootstrapTokenForm) {
+    verifyMutation.mutate(data.token.trim(), {
+      onSuccess: (res) => {
+        setSessionToken(res.session_token)
+        setSessionExpiresAt(res.expires_at)
+        setCurrentStep(1)
+        void navigate({ to: '/setup/mode' })
+      },
+    })
+  }
+
+  return (
+    <Form {...form}>
+      <form
+        onSubmit={form.handleSubmit(handleFormSubmit)}
+        className="space-y-4"
+      >
+        <PageMeta title="Setup Token" />
+        <div className="space-y-1">
+          <h2 className="text-lg font-medium">Bootstrap Token</h2>
+          <p className="text-sm text-muted-foreground">
+            Enter the bootstrap token generated by the{' '}
+            <code className="bg-muted px-1 py-0.5 text-xs">
+              oore setup token
+            </code>{' '}
+            CLI command.
+          </p>
+        </div>
+
+        <FormField
+          control={form.control}
+          name="token"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Token</FormLabel>
+              <FormControl>
+                <Input
+                  type="password"
+                  placeholder="Paste your bootstrap token"
+                  disabled={verifyMutation.isPending}
+                  autoFocus
+                  {...field}
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        {errorMessage ? (
+          <Alert variant="destructive">
+            <AlertTitle>Verification failed</AlertTitle>
+            <AlertDescription>{errorMessage}</AlertDescription>
+          </Alert>
+        ) : null}
+
+        <Button
+          type="submit"
+          disabled={verifyMutation.isPending}
+          className="w-full"
+        >
+          {verifyMutation.isPending ? 'Verifying...' : 'Verify token'}
+        </Button>
+      </form>
+    </Form>
+  )
+}
